Comment
I'm a little concerned with the proliferation of "Actions" files. The intent is to have one "Actions" file per module... and actions are not utilities. Utilities should be in their own file(s). 

assignment-inputs/Actions should define all the actions for this "sub-module".  We shouldn't have "Actions" in the various input type sub directries...